+++ NEWS.md
@@ -1,5 +1,7 @@
 # batchtools 0.9.1
 * Deactivated swap for clusterFunctionsDocker.
+* `findExperiments` now has two additional arguments to match for regular expressions.
+  The possibility to prefix a string with "~" to enable regular expression matching has been removed.
 
 # batchtools 0.9.0